+++ src/eca/llm_providers/google.clj
@@ -2,24 +2,12 @@
   (:require
    [clojure.string :as string]
    [eca.config :as config]
-   [eca.features.login :as f.login]
-   [eca.shared :refer [multi-str]]))
+   [eca.features.login :as f.login]))
 
 (defmethod f.login/login-step ["google" :login/start] [{:keys [db* chat-id provider send-msg!]}]
   (swap! db* assoc-in [:chats chat-id :login-provider] provider)
-  (swap! db* assoc-in [:auth provider] {:step :login/waiting-login-method})
-  (send-msg! (multi-str "Now, inform the login method:"
-                        ""
-                        "manual: Manually enter API Key")))
-
-(defmethod f.login/login-step ["google" :login/waiting-login-method] [{:keys [db* input provider send-msg!]}]
-  (case input
-    "manual"
-    (do
-      (swap! db* assoc-in [:auth provider] {:step :login/waiting-api-key
-                                            :mode :manual})
-      (send-msg! "Paste your google API Key"))
-    (send-msg! (format "Unknown login method '%s'. Inform one of the options: max, console, manual" input))))
+  (swap! db* assoc-in [:auth provider] {:step :login/waiting-api-key})
+  (send-msg! "Paste your google API Key"))
 
 (defmethod f.login/login-step ["google" :login/waiting-api-key] [{:keys [db* input provider send-msg!] :as ctx}]
   (if (not (string/blank? input))